@@ -49,60 +55,63 @@
     "protractor:delay": "sleep 3 && npm run protractor",
     "protractor:live": "protractor --elementExplorer",
     "rimraf": "rimraf",
-    "server:dev:hmr": "npm run server:dev -- --inline --hot",
+    "server:dev:hmr": "npm run server:dev -- --hotOnly",
+    "server:aot:dev": "cross-env BUILD_AOT=1 npm run server:dev",
     "server:dev": "npm run webpack-dev-server -- --config config/webpack.dev.js --open --progress --profile --watch --content-base src/",
     "server:prod": "http-server dist -c-1 --cors",
     "server:prod:ci": "http-server dist -p 3000 -c-1 --cors",
     "server": "npm run server:dev",
     "start:hmr": "npm run server:dev:hmr",
     "start": "npm run server:dev",
+    "start:aot": "npm run server:aot:dev",
     "test": "npm run lint && karma start",
     "tslint": "tslint",
     "typedoc": "typedoc",
     "version": "npm run build",
     "watch:dev:hmr": "npm run watch:dev -- --hot",
     "watch:dev": "npm run build:dev -- --watch",
+    "watch:aot:dev": "npm run build:aot:dev -- --watch",
     "watch:prod": "npm run build:prod -- --watch",
+    "watch:aot:prod": "npm run build:aot:prod -- --watch",
     "watch:test": "npm run test -- --auto-watch --no-single-run",
     "watch": "npm run watch:dev",
-    "webdriver-manager": "webdriver-manager",
-    "webdriver:start": "npm run webdriver-manager start",
-    "webdriver:update": "webdriver-manager update",
+    "webdriver-manager": "node ./node_modules/protractor/bin/webdriver-manager",
+    "webdriver:start": "node ./node_modules/protractor/bin/webdriver-manager start",
+    "webdriver:update": "node ./node_modules/protractor/bin/webdriver-manager update",
     "webpack-dev-server": "webpack-dashboard -- node --max_old_space_size=4096 node_modules/webpack-dev-server/bin/webpack-dev-server.js",
     "webpack": "node --max_old_space_size=4096 node_modules/webpack/bin/webpack.js"
   },
   "dependencies": {
-    "@angular/animations": "^4.4.4",
-    "@angular/cdk": "^2.0.0-beta.12",
-    "@angular/common": "^4.4.4",
-    "@angular/compiler": "^4.4.4",
-    "@angular/core": "^4.4.4",
-    "@angular/forms": "^4.4.4",
-    "@angular/http": "^4.4.4",
-    "@angular/material": "^2.0.0-beta.12",
-    "@angular/platform-browser": "^4.4.4",
-    "@angular/platform-browser-dynamic": "^4.4.4",
-    "@angular/platform-server": "^4.4.4",
-    "@angular/router": "^4.4.4",
-    "@angularclass/hmr": "~1.2.2",
-    "@angularclass/hmr-loader": "^3.0.4",
+    "@angular/animations": "~5.0.1",
+    "@angular/common": "~5.0.1",
+    "@angular/compiler": "~5.0.1",
+    "@angular/core": "~5.0.1",
+    "@angular/forms": "~5.0.1",
+    "@angular/http": "~5.0.1",
+    "@angular/platform-browser": "~5.0.1",
+    "@angular/platform-browser-dynamic": "~5.0.1",
+    "@angular/platform-server": "~5.0.1",
+    "@angular/router": "~5.0.1",
+    "core-js": "^2.5.1",
+    "http-server": "^0.9.0",
+    "ie-shim": "^0.1.0",
+    "reflect-metadata": "^0.1.10",
+    "rxjs": "^5.5.2",
+    "zone.js": "~0.8.18"
     "@ngx-translate/core": "^7.1.0",
     "@ngx-translate/http-loader": "^0.1.0",
     "aws-sdk": "^2.96.0",
-    "core-js": "^2.4.1",
-    "http-server": "^0.9.0",
-    "ie-shim": "^0.1.0",
     "moment": "^2.18.1",
     "pouchdb": "^6.3.4",
     "pouchdb-adapter-localstorage": "^6.3.4",
     "pouchdb-browser": "^6.3.4",
-    "reflect-metadata": "^0.1.10",
-    "rxjs": "~5.0.2",
-    "web-animations-js": "^2.3.1",
-    "zone.js": "0.8.14"
+    "web-animations-js": "^2.3.1"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^4.4.4",
+    "@angular-devkit/build-optimizer": "^0.0.32",
+    "@angular/compiler-cli": "~5.0.1",
+    "@compodoc/compodoc": "^1.0.0-beta.15",
+    "@ngtools/webpack": "^1.8.0",
     "@types/hammerjs": "^2.0.34",
     "@types/jasmine": "2.5.45",
     "@types/node": "^7.0.39",
@@ -160,21 +167,22 @@
     "tslint": "~4.5.1",
     "tslint-loader": "^3.5.2",
     "typedoc": "^0.7.1",
-    "typescript": "~2.3.4",
+    "typescript": "~2.5.2",
+    "uglifyjs-webpack-plugin": "^1.0.1",
     "url-loader": "^0.5.8",
-    "webpack": "^3.6.0",
-    "webpack-dashboard": "^0.4.0",
+    "webpack": "^3.8.1",
     "webpack-dev-middleware": "^1.10.1",
-    "webpack-dev-server": "~2.4.2",
+    "webpack-dev-server": "~2.7.1",
     "webpack-dll-bundles-plugin": "^1.0.0-beta.5",
     "webpack-merge": "~4.1.0"
+    "webpack-dashboard": "^0.4.0",
   },
   "repository": {
     "type": "git",
-    "url": "https://github.com/AngularClass/angular-starter.git"
+    "url": "https://github.com/gdi2290/angular-starter.git"
   },
   "bugs": {
-    "url": "https://github.com/AngularClass/angular-starter/issues"
+    "url": "https://github.com/gdi2290/angular-starter/issues"
   },
   "engines": {
     "node": ">= 4.2.1",
