@@ -9,7 +9,7 @@ module.exports = function (config) {
 
     /**
      * Base path that will be used to resolve all patterns (e.g. files, exclude).
-     */
+    */
     basePath: '',
 
     /**
@@ -21,7 +21,7 @@ module.exports = function (config) {
 
     /**
      * List of files to exclude.
-     */
+    */
     exclude: [],
 
     client: {
