@@ -6,7 +6,7 @@ require('ts-node/register');
 var helpers = require('./helpers');
 
 exports.config = {
-  baseUrl: 'http://localhost:8888/',
+  baseUrl: 'http://localhost:3000/',
 
   /**
    * Use `npm run e2e`
