@@ -3,37 +3,17 @@
  */
 
 const helpers = require('./helpers');
+const buildUtils = require('./build-utils');
 const webpackMerge = require('webpack-merge'); // used to merge webpack configs
-// const webpackMergeDll = webpackMerge.strategy({plugins: 'replace'});
 const commonConfig = require('./webpack.common.js'); // the settings that are common to prod and dev
 
 /**
  * Webpack Plugins
  */
-const AddAssetHtmlPlugin = require('add-asset-html-webpack-plugin');
-const DefinePlugin = require('webpack/lib/DefinePlugin');
-const NamedModulesPlugin = require('webpack/lib/NamedModulesPlugin');
 const LoaderOptionsPlugin = require('webpack/lib/LoaderOptionsPlugin');
+const NamedModulesPlugin = require('webpack/lib/NamedModulesPlugin');
+const EvalSourceMapDevToolPlugin = require('webpack/lib/EvalSourceMapDevToolPlugin');
 const DashboardPlugin = require('webpack-dashboard/plugin');
-
-/**
- * Webpack Constants
- */
-const ENV = process.env.ENV = process.env.NODE_ENV = 'development';
-const HOST = process.env.HOST || 'localhost';
-const PORT = process.env.PORT || 3000;
-const PUBLIC = process.env.PUBLIC_DEV || undefined;
-const HMR = helpers.hasProcessFlag('hot');
-const METADATA = webpackMerge(commonConfig({env: ENV}).metadata, {
-  host: HOST,
-  port: PORT,
-  public: PUBLIC,
-  ENV: ENV,
-  HMR: HMR
-});
-
-
-// const DllBundlesPlugin = require('webpack-dll-bundles-plugin').DllBundlesPlugin;
 
 /**
  * Webpack configuration
@@ -41,16 +21,19 @@ const METADATA = webpackMerge(commonConfig({env: ENV}).metadata, {
  * See: http://webpack.github.io/docs/configuration.html#cli
  */
 module.exports = function (options) {
-  return webpackMerge(commonConfig({env: ENV}), {
+  const ENV = process.env.ENV = process.env.NODE_ENV = 'development';
+  const HOST = process.env.HOST || 'localhost';
+  const PORT = process.env.PORT || 3000;
 
-    /**
-     * Developer tool to enhance debugging
-     *
-     * See: http://webpack.github.io/docs/configuration.html#devtool
-     * See: https://github.com/webpack/docs/wiki/build-performance#sourcemaps
-     */
-    devtool: 'cheap-module-source-map',
+  const METADATA = Object.assign({}, buildUtils.DEFAULT_METADATA, {
+    host: HOST,
+    port: PORT,
+    ENV: ENV,
+    HMR: helpers.hasProcessFlag('hot'),
+    PUBLIC: process.env.PUBLIC_DEV || HOST + ':' + PORT
+  });
 
+  return webpackMerge(commonConfig({ env: ENV, metadata: METADATA  }), {
     /**
      * Options affecting the output of the compilation.
      *
@@ -123,72 +106,10 @@ module.exports = function (options) {
     },
 
     plugins: [
-
-      /**
-       * Plugin: DefinePlugin
-       * Description: Define free variables.
-       * Useful for having development builds with debug logging or adding global constants.
-       *
-       * Environment helpers
-       *
-       * See: https://webpack.github.io/docs/list-of-plugins.html#defineplugin
-       *
-       * NOTE: when adding more properties, make sure you include them in custom-typings.d.ts
-       */
-      new DefinePlugin({
-        'ENV': JSON.stringify(METADATA.ENV),
-        'HMR': METADATA.HMR,
-        'process.env': {
-          'ENV': JSON.stringify(METADATA.ENV),
-          'NODE_ENV': JSON.stringify(METADATA.ENV),
-          'HMR': METADATA.HMR,
-        }
+      new EvalSourceMapDevToolPlugin({
+        moduleFilenameTemplate: '[resource-path]',
+        sourceRoot: 'webpack:///'
       }),
-
-      // new DllBundlesPlugin({
-      //   bundles: {
-      //     polyfills: [
-      //       'core-js',
-      //       {
-      //         name: 'zone.js',
-      //         path: 'zone.js/dist/zone.js'
-      //       },
-      //       {
-      //         name: 'zone.js',
-      //         path: 'zone.js/dist/long-stack-trace-zone.js'
-      //       },
-      //     ],
-      //     vendor: [
-      //       '@angular/platform-browser',
-      //       '@angular/platform-browser-dynamic',
-      //       '@angular/core',
-      //       '@angular/common',
-      //       '@angular/forms',
-      //       '@angular/http',
-      //       '@angular/router',
-      //       '@angularclass/hmr',
-      //       'rxjs',
-      //     ]
-      //   },
-      //   dllDir: helpers.root('dll'),
-      //   webpackConfig: webpackMergeDll(commonConfig({env: ENV}), {
-      //     devtool: 'cheap-module-source-map',
-      //     plugins: []
-      //   })
-      // }),
-
-      /**
-       * Plugin: AddAssetHtmlPlugin
-       * Description: Adds the given JS or CSS file to the files
-       * Webpack knows about, and put it into the list of assets
-       * html-webpack-plugin injects into the generated html.
-       *
-       * See: https://github.com/SimenB/add-asset-html-webpack-plugin
-       */
-      // new AddAssetHtmlPlugin([
-      //   { filepath: helpers.root(`dll/${DllBundlesPlugin.resolveFile('polyfills')}`) },
-      //   { filepath: helpers.root(`dll/${DllBundlesPlugin.resolveFile('vendor')}`) }
-      // ]),
 
       /**
        * Plugin: NamedModulesPlugin (experimental)
@@ -196,7 +117,7 @@ module.exports = function (options) {
        *
        * See: https://github.com/webpack/webpack/commit/a04ffb928365b19feb75087c63f13cadfc08e1eb
        */
-      // new NamedModulesPlugin(),
+      new NamedModulesPlugin(),
 
       /**
        * Plugin LoaderOptionsPlugin (experimental)
@@ -205,9 +126,7 @@ module.exports = function (options) {
        */
       new LoaderOptionsPlugin({
         debug: true,
-        options: {
-
-        }
+        options: { }
       }),
 
       new DashboardPlugin()
@@ -224,9 +143,9 @@ module.exports = function (options) {
      */
     devServer: {
       port: METADATA.port,
-      host: '0.0.0.0',
-      disableHostCheck: true,
-      public: METADATA.public,
+      host: METADATA.host,
+      hot: METADATA.HMR,
+      public: METADATA.PUBLIC,
       historyApiFallback: true,
       watchOptions: {
         // if you're using Docker you may need this
@@ -263,4 +182,4 @@ module.exports = function (options) {
     }
 
   });
-};
+}
