@@ -3,15 +3,15 @@
  * When including 3rd party modules you also need to include the type definition for the module
  * if they don't provide one within the module. You can try to install it with @types
 
- npm install @types/node
- npm install @types/lodash
+npm install @types/node
+npm install @types/lodash
 
  * If you can't find the type definition in the registry we can make an ambient/global definition in
  * this file for now. For example
 
- declare module 'my-module' {
+declare module 'my-module' {
  export function doesSomething(value: string): string;
- }
+}
 
  * If you are using a CommonJS module that is using module.exports then you will have to write your
  * types using export = yourObjectOrFunction with a namespace above it
@@ -81,15 +81,15 @@ type FactoryPromise = () => Promise<any>;
 
 type AsyncRoutes = {
   [component: string]: Es6PromiseLoader |
-    Function |
-    FactoryEs6PromiseLoader |
-    FactoryPromise;
+                               Function |
+                FactoryEs6PromiseLoader |
+                         FactoryPromise ;
 };
 
 type IdleCallbacks = Es6PromiseLoader |
-  Function |
-  FactoryEs6PromiseLoader |
-  FactoryPromise ;
+                             Function |
+              FactoryEs6PromiseLoader |
+                       FactoryPromise ;
 
 interface WebpackModule {
   hot: {
@@ -108,14 +108,14 @@ interface WebpackModule {
 }
 
 interface WebpackRequire {
-  (id: string): any;
-  (paths: string[], callback: (...modules: any[]) => void): void;
-  ensure(ids: string[], callback: (req: WebpackRequire) => void, chunkName?: string): void;
-  context(directory: string, useSubDirectories?: boolean, regExp?: RegExp): WebpackContext;
+    (id: string): any;
+    (paths: string[], callback: (...modules: any[]) => void): void;
+    ensure(ids: string[], callback: (req: WebpackRequire) => void, chunkName?: string): void;
+    context(directory: string, useSubDirectories?: boolean, regExp?: RegExp): WebpackContext;
 }
 
 interface WebpackContext extends WebpackRequire {
-  keys(): string[];
+    keys(): string[];
 }
 
 interface ErrorStackTraceLimit {
@@ -125,8 +125,6 @@ interface ErrorStackTraceLimit {
 // Extend typings
 interface NodeRequire extends WebpackRequire {}
 interface ErrorConstructor extends ErrorStackTraceLimit {}
-interface NodeRequireFunction extends Es6PromiseLoader {
-}
+interface NodeRequireFunction extends Es6PromiseLoader  {}
 interface NodeModule extends WebpackModule {}
-interface Global extends GlobalEnvironment {
-}
+interface Global extends GlobalEnvironment  {}
